Se implementa guardar, editar y eliminar Cita
· Se implementa el guardado, modificaicón y eliminación de datos enla tabla Cita
· También se implementa un poco de estandarización en la capa de servicios con los datos de tipo Long

+++ src/main/java/cl/petit/api/controllers/MascotaController.java
@@ -58,7 +58,7 @@ public class MascotaController {
 
         Map<String, Object> result = new HashMap<String,Object>();
         if(enviar){
-            ArrayList<MascotaDTO> mascotas = this.mascotaService.obtenerConPagina(pagina);
+            ArrayList<MascotaDTO> mascotas = this.mascotaService.obtenerConPagina(pagina.longValue());
             if(mascotas.size()>0){
                 result.put("result",true);
                 result.put("mascotas",mascotas);
